+++ ya/options.go
@@ -1,0 +1,5 @@
+package ya
+
+type DownloadOptions struct {
+	SkipCover bool
+}